requirements.txt
pyjwt
pygobject